feat: functions
Functions are kind of stereotypical, the have inputs with a type and if
they have return values, these types need to be declared as well.

As Rust uses the `;` to end lines the only surprise is how to define the
return value. Either use the explicit form of `return foo;` or just by
using `foo` without either of return and a semicolon.