feat: first contact with the infamous borrow checker
For a language that isn't garbage collected but has to make assumptions
abot variable lifetimes during compile time without explicit calls of
free() or malloc() the borrow checker comes into play.

Move semantics are a way for the compiler to track who owns the
data/memory of a variable at a given time. It can be complex and it took
a while to gain a first understanding.